Fixes (FACILITIES-97)
Because of the way the event wiring facility was using to resolve a component by passing an empty CreationContext
to the handler any dependencies on a generic service would fail because the call to CloneParentProperties expects the
context to have a handler associated with it.

I change the facility to use Kernel.Resolve instead since it looks like the ResolveComponent code was doing the same thing
as the facility was trying to do instead.


git-svn-id: http://svn.castleproject.org:8080/svn/castle/trunk/InversionOfControl@4721 73e77b4c-caa6-f847-a29a-24ab75ae54b6